KeyWorkExtraction_Python/requirements.txt
# Usage: pip install -r requirements.txt

# Base ------------------------------------------------------------------------
gitpython>=3.1.30
matplotlib>=3.3
numpy>=1.22.2
requests>=2.31.0
scipy>=1.4.1
thop>=0.1.1  # FLOPs computation

# progress percentage %
tqdm>=4.64.0

# object prediction -------------------------------------------------
# ultralytics>=8.0.147

# TF-IDF
scikit-learn

# LDA Latent Dirichlet Allocation
pyLDAvis==2.1.2

# html parse library  -------------------------------------------------
beautifulsoup4
# pohan
flask_restful
flask_cors
# protobuf<=3.20.1  # https://github.com/ultralytics/yolov5/issues/8012

# Logging ---------------------------------------------------------------------
# tensorboard>=2.4.1
# clearml>=1.2.0
# comet

# Plotting (drawing) --------------------------------------------------------------------
# pandas>=1.1.4
# seaborn>=0.11.0

# comet_ml comet_ml 是一個用於機器學習實驗管理和協作的工具。它提供了許多功能，包括實驗跟蹤、超參數調整、模型性能比較、圖表視覺化等。使用 comet_ml 可以幫助您更有效地管理機器學習項目，追蹤實驗結果，並與團隊成員共享和協作。
# 您可以使用 comet_ml 在不同的機器學習項目中記錄和比較實驗結果，並通過其提供的視覺化工具更好地了解模型的表現。

comet_ml

# Export ----------------------------------------------------------------------
# coremltools>=6.0  # CoreML export
# onnx>=1.10.0  # ONNX export
# onnx-simplifier>=0.4.1  # ONNX simplifier
# nvidia-pyindex  # TensorRT export
# nvidia-tensorrt  # TensorRT export
# scikit-learn<=1.1.2  # CoreML quantization
# tensorflow>=2.4.0  # TF exports (-cpu, -aarch64, -macos)
# tensorflowjs>=3.9.0  # TF.js export
# openvino-dev>=2023.0  # OpenVINO export

# Deploy ----------------------------------------------------------------------
# setuptools>=65.5.1 # Snyk vulnerability fix
# tritonclient[all]~=2.24.0

# Extras ----------------------------------------------------------------------
# ipython  # interactive notebook
# mss  # screenshots
# albumentations>=1.0.3
# pycocotools>=2.0.6  # COCO mAP
